refactor(events): extract host-action endpoints into _event_host_actions.py
Moves get_event_stats, get_text_recipients, and set_attendance (plus their
private helpers) out of _event_rsvps.py into a new _event_host_actions.py.
No behavior change — pure code move ahead of the upcoming host rsvp-management
feature (Issue 872).
