+++ backend/openapi_schema.json
@@ -10026,7 +10026,7 @@
     },
     "/api/community/events/{event_id}/rsvps/{user_id}/attendance/": {
       "post": {
-        "operationId": "community__event_rsvps_set_attendance",
+        "operationId": "community__event_host_actions_set_attendance",
         "parameters": [
           {
             "in": "path",
@@ -10124,7 +10124,7 @@
     },
     "/api/community/events/{event_id}/stats/": {
       "get": {
-        "operationId": "community__event_rsvps_get_event_stats",
+        "operationId": "community__event_host_actions_get_event_stats",
         "parameters": [
           {
             "in": "path",
@@ -10182,7 +10182,7 @@
     },
     "/api/community/events/{event_id}/text-recipients/": {
       "get": {
-        "operationId": "community__event_rsvps_get_text_recipients",
+        "operationId": "community__event_host_actions_get_text_recipients",
         "parameters": [
           {
             "in": "path",
